Solution for 138. Copy List with Random Pointer //Creating a new linked list and adding the nodes one by one will cause a problem, which is, where should the random pointer point if the node it is pointing at does not exist yet.
To fix this problem, we can create a hashmap where the old nodes maps to cloned nodes with values of the nodes only respectively.

Then, by using the mapping in the hashmaps, we can set the next pointer and the random pointer of the node in the value of hashmap by looking up the key in the hashmap/ original node.

Finally, we use the input linked list to find the generated linked list